+++ Assets/Scripts/Voxel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ba646c4a964204143a3720730709e4f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
